+++ generator/template/config/storybook/addons.js
@@ -1,2 +1,3 @@
+/* eslint-disable import/no-extraneous-dependencies */
 import '@storybook/addon-actions/register'
 import '@storybook/addon-links/register'