+++ generator/template/config/storybook/config.js
@@ -1,3 +1,4 @@
+/* eslint-disable import/no-extraneous-dependencies */
 import { configure } from '@storybook/vue'
 
 const req = require.context('../../src/stories', true, /.stories.js$/)